# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Chemnitz open data map for Database \u0026 Web Technologies",
        "title": "DWT-Chemnitz-map",
        "contact": {
            "name": "Vitalii Popov",
            "url": "https://www.linkedin.com/in/mskVitalii/",
            "email": "msk.vitaly@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:80",
    "basePath": "/",
    "paths": {
        "/api/v1/claims": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fetches and returns the claims of the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Retrieves user claims",
                "responses": {
                    "200": {
                        "description": "Successful response with user claims",
                        "schema": {
                            "$ref": "#/definitions/dto.GetUserClaimsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.UnauthorizedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.UnauthorizedResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/login": {
            "post": {
                "description": "Authenticates a user with email and password, and returns a JWT token if the credentials are valid.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Authenticates a user and returns a JWT token",
                "parameters": [
                    {
                        "description": "Login request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessLoginResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.UnauthorizedResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "description": "Invalidates the JWT token for the user by adding it to the blacklist and removes the JWT cookie.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Logs out a user and invalidates the JWT token",
                "responses": {
                    "200": {
                        "description": "Successful logout",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessLogoutResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/lib.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/refresh_token": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generates a new authentication token using the refresh token provided.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refreshes the authentication token",
                "responses": {
                    "200": {
                        "description": "Successful response with new token",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessRefreshTokenResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.UnauthorizedResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.GetUserClaimsResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "userType": {
                    "type": "string"
                }
            }
        },
...
        "dto.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "dto.SuccessLoginResponse": {
            "type": "object",
            "properties": {
                "expire": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.SuccessLogoutResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.SuccessRefreshTokenResponse": {
            "type": "object",
            "properties": {
                "expire": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.UnauthorizedResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "lib.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}